refactor(ol-params): move params types out of lib
Keep lib.rs as the crate surface and move OLParams, OLGenesisParams, and OLRuntimeParams into params.rs. Trim the moved tests to behavior owned by the split params shape.
